r.in.lidar: use segment lib to enable different resolution for output
Using Segment library is easier than implementing another segmentation
similar to the existing one (based on rows); the impl. would also require
way of getting the right row/col based on corrdinates.

Leaving there the raster row based approach for cases when resolution
of output is not set. It is faster since it does not create the segments.

This fixes segfault and fatal when the resolution option was lower or
higher than the current one (introduced in r66094).

Deliberately not using the old formatting style.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@66151 15284696-431f-4ddb-bdfa-cd5b030d7da7