doc: keyword name is miscellaneous, so use it for the page as well
Parser (print_escaped_for_html_keywords()) just writes the keyword names
but the file was called misc and the keyword miscellaneous.

This makes the URL or index name harder to type manually
but it makes (first) keyword links work (for m.* modules)
and it makes the titles and link texts Miscellaneous instead
of Misc.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@69588 15284696-431f-4ddb-bdfa-cd5b030d7da7