doc: Python 3 compatibility (see #2708)
Decode file content on input in 3, leave as in in 2.
Call constructor of parent class (Python 3 version needs to init a var).
Call less efficient items() in Python 2 (same code in both, although slower in 2, ok here).
This avoids compile (mkhtml) errors in most of the directories with Python 3.5.2 as python.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@71849 15284696-431f-4ddb-bdfa-cd5b030d7da7