init: add --tmp-location (based on patch from radeknovotny94 in #3537)
Creates a location in program's tmp dir which is deleted at the end.
Meant for use with --exec. Possible to use without it as well (e.g. with GUI).
CRS needs to be specified using -c.
Does not write to a rc file to user's settings at the end.
Based on grass-tmp-loc-2.diff from Radek Novotny submitted in #3537.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@72790 15284696-431f-4ddb-bdfa-cd5b030d7da7