init: remove basically dead code to clean mapset .tmp in default startup (see #3635)
The default (fallback) startup contained code which was probably from the
original Python rewrite (r37863) and maybe originated from workaround in r21048.
The clean_temp tool is called and takes care of cleaning, so this should not
be neccessary, not even for GUI (mentioned in the comment). GUI is not using
mapset temporary directory since r73334 (#3637).


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@73337 15284696-431f-4ddb-bdfa-cd5b030d7da7