init: use MAPSET_PATH instead of LOCATION for full path to current mapset variable (see #2681)
Changes the (local) shell variable used to refer to full path to the
current mapset from LOCATION to MAPSET_PATH. The name LOCATION was
a lagacy name (see also LOCATION_NAME) reintroduced in r60216
(not mentioned in the commit message, used locally like in the Python prompt).
The LOCATION is kept for backwards compatibility and can be safely removed
for the next major release (8.0).


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@73352 15284696-431f-4ddb-bdfa-cd5b030d7da7