v.in.pdal: use PDAL streaming instead of PointView
The memory consumption no longer depends on the point cloud size.
Instead, it is fixed internally. Higher number of points to store
does not make it faster, so keeping it small.

What was a loop is now a lambda used as a callback in the StreamCallbackFilter
which filters and writes the vector points.

The limit on category number (GV_CAT_MAX) is enforced in the reader,
so we don't check it when reading (if cat is requested).


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@74484 15284696-431f-4ddb-bdfa-cd5b030d7da7